Add configurable context window size
It seems that context window of 200 was way too small, I saw quite many
`_________id` style renamess in the issues. Using a bigger context
window should fix this issue, and mmaking the window configurable
should help to debug any future issues.

This value is a balance between giing enough context for the LLM to
infer the meaning correctly, and overwhelming the LLM with too much
noise.